change mapping to get, add find by id feature

+++ src/main/java/com/medbis/repository/DiseaseRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface DiseaseRepository extends JpaRepository<Disease, Long> {
+public interface DiseaseRepository extends JpaRepository<Disease, Integer> {
 
 }
